requirements.txt
slack-bolt==1.18.1
slack-sdk==3.27.1
openai>=1.0.0
python-dotenv==1.0.1
